+++ types/protocol.d.ts
@@ -4540,6 +4540,20 @@ export namespace Protocol {
         }
 
         /**
+         * CSS style coming from animations with the name of the animation.
+         */
+        export interface CSSAnimationStyle {
+            /**
+             * The name of the animation.
+             */
+            name?: string;
+            /**
+             * The style coming from the animation.
+             */
+            style: CSSStyle;
+        }
+
+        /**
          * Inherited CSS rule collection from ancestor node.
          */
         export interface InheritedStyleEntry {
@@ -4551,6 +4565,20 @@ export namespace Protocol {
              * Matches of CSS rules matching the ancestor node in the style inheritance chain.
              */
             matchedCSSRules: RuleMatch[];
+        }
+
+        /**
+         * Inherited CSS style collection for animated styles from ancestor node.
+         */
+        export interface InheritedAnimatedStyleEntry {
+            /**
+             * Styles coming from the animations of the ancestor, if any, in the style inheritance chain.
+             */
+            animationStyles?: CSSAnimationStyle[];
+            /**
+             * The style coming from the transitions of the ancestor, if any, in the style inheritance chain.
+             */
+            transitionsStyle?: CSSStyle;
         }
 
         /**
@@ -5526,6 +5554,26 @@ export namespace Protocol {
              * Attribute-defined element style (e.g. resulting from "width=20 height=100%").
              */
             attributesStyle?: CSSStyle;
+        }
+
+        export interface GetAnimatedStylesForNodeRequest {
+            nodeId: DOM.NodeId;
+        }
+
+        export interface GetAnimatedStylesForNodeResponse {
+            /**
+             * Styles coming from animations.
+             */
+            animationStyles?: CSSAnimationStyle[];
+            /**
+             * Style coming from transitions.
+             */
+            transitionsStyle?: CSSStyle;
+            /**
+             * Inherited style entries for animationsStyle and transitionsStyle from
+             * the inheritance chain of the element.
+             */
+            inherited?: InheritedAnimatedStyleEntry[];
         }
 
         export interface GetMatchedStylesForNodeRequest {